Move MyError into its own file
